+++ crates/translator_new_rml/src/translator/source/kind/mod.rs
@@ -1,3 +1,4 @@
+pub mod csvw_source;
 pub mod kafka_source;
 pub mod rdb_source;
 pub mod tcp_source;